Revert "doc/Makefile: drop doc-diff worktree and temporary files on "make clean""
This reverts commit 0bcf6b3a9e8d3584b0bf41800246eda41723272a, which
started to require that we have an executable git available in order
to say "make clean", which gives us a chicken-and-egg problem.

Having to have Git installed, or be in a repository, in order to be
able to run an optional "doc-diff" tool is fine.  Requiring either
in order to run "make clean" is a different story.

Reported by Jonathan Nieder <jrnieder@gmail.com>.